Alternative radiotherapy models
Implemented two alternative radiotherapy models. The first approach decreases the proliferation rate with each dose of radiotherapy. The assumption is that the population of actively proliferating cells is changing over time. This is  controlled by assigning "alpha_proliferation" in the radiotherapy specification.   The second approach increases the death rate due to radiotherapy with each dose of radiotherapy. The assumption here is that with each dose of radiotherapy the tumor will begin to die off.  This is controlled by setting "k_d" to a non-zero value, and setting "alpha_death" in the radiotherapy specification.  For model calibration, set one of those to a fixed value and then fit for the other.

Both of these models relate to Eq. 7 and 8 as reported in https://www.mdpi.com/2072-6694/13/8/1765